@@ -1,5 +1,6 @@
 package com.example.ssussemble;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -7,6 +8,12 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -54,9 +61,7 @@ public class ChatAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
     @Override
     public int getItemViewType(int position) {
         ChatData chat = chatList.get(position);
-
-        String nickName = MainActivity.Login_id.substring(0, MainActivity.Login_id.indexOf("@"));
-        if (chat.getUserName().equals(nickName)) {
+        if (chat.getUserName().equals(MainActivity.Login_id)) {
             return VIEW_TYPE_MINE;
         } else {
             return VIEW_TYPE_OTHER;
@@ -105,8 +110,19 @@ public class ChatAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
         }
 
         void bind(ChatData chat) {
+            // 메시지 설정
             messageText.setText(chat.getMessage());
-            userNameText.setText(chat.getUserName());
+
+            // 이메일 기반으로 displayName 가져오기
+            fetchDisplayName(chat.getUserName(), displayName -> {
+                if (displayName != null && !displayName.isEmpty()) {
+                    userNameText.setText(displayName);
+                } else {
+                    userNameText.setText("Unknown User");
+                }
+            });
+
+            // 타임스탬프 설정
             if (chat.getTimestamp() != 0) {
                 SimpleDateFormat sdf = new SimpleDateFormat("HH:mm", Locale.getDefault());
                 sdf.setTimeZone(TimeZone.getTimeZone("Asia/Seoul"));
@@ -114,4 +130,36 @@ public class ChatAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
             }
         }
     }
+
+
+    private static void fetchDisplayName(String userName, OnDisplayNameFetchedListener listener) {
+        DatabaseReference usersRef = FirebaseDatabase.getInstance().getReference("users");
+
+        usersRef.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                for (DataSnapshot userSnapshot : snapshot.getChildren()) {
+                    UserData userData = userSnapshot.getValue(UserData.class);
+                    if (userData != null && userName.equals(userData.getEmail())) {
+                        listener.onFetched(userData.getDisplayName());
+                        return;
+                    }
+                }
+                // 일치하는 사용자 없음
+                listener.onFetched("Unknown User");
+            }
+
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+                Log.e("Firebase", "Failed to fetch displayName", error.toException());
+                listener.onFetched("Unknown User");
+            }
+        });
+    }
+
+    // 콜백 인터페이스 정의
+    public interface OnDisplayNameFetchedListener {
+        void onFetched(String displayName);
+    }
+
 }
